refactor: 카드 종류를 결정하는 함수 이름과 커스텀 훅 이름 수정
